requirements copy.txt
Django==4.2.15
django-cors-headers==4.3.1

djangorestframework==3.15.2
django-filter==23.1
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
daphne==4.0.0
# psycopg2==2.9.9
psycopg2-binary
drf-yasg==1.21.7
pillow==10.3.0

google-auth==2.26.1
requests
pandas
phonenumbers==8.13.43

django-admin-interface==0.28.5
openpyxl
whitenoise==6.7.0


# https://whimsical.com/trekking-site-EKtejmjwcztrsy7KnNKe2F